更新  INDEX 與 機械座標 建立

+++ WLS3200Gen2.Model/Recipe/Die.cs
@@ -10,11 +10,19 @@ namespace YuanliCore.Data
 {
     public class Die
     {
+        /// <summary>
+        /// 
+        /// </summary>
         public int IndexX { get; set; }
         public int IndexY { get; set; }
-
-        public int PosX { get; set; }
-        public int PosY { get; set; }
+        /// <summary>
+        /// 機械座標x
+        /// </summary>
+        public double PosX { get; set; }
+        /// <summary>
+        /// 機械座標Y
+        /// </summary>
+        public double PosY { get; set; }
 
         public Size DieSize { get; set; }
     }
